+++ ArchiWindRevitAddIn/Api/Models/BillingPlan.cs
@@ -49,7 +49,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static global::ArchiwindRevitAddIn.Api.Models.BillingPlan CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.Models.BillingPlan();
         }
         /// <summary>
@@ -72,7 +72,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="writer">Serialization writer to use to serialize this model</param>
         public virtual void Serialize(ISerializationWriter writer)
         {
-            _ = writer ?? throw new ArgumentNullException(nameof(writer));
+            if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_detailed_credits>("detailed_credits", DetailedCredits);
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_draft_credits>("draft_credits", DraftCredits);
             writer.WriteDateValue("expires_on", ExpiresOn);
@@ -102,7 +102,7 @@ namespace ArchiwindRevitAddIn.Api.Models
             /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
             public static global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_detailed_credits CreateFromDiscriminatorValue(IParseNode parseNode)
             {
-                _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+                if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
                 var mappingValue = parseNode.GetChildNode("")?.GetStringValue();
                 var result = new global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_detailed_credits();
                 if(parseNode.GetIntValue() is int integerValue)
@@ -129,7 +129,7 @@ namespace ArchiwindRevitAddIn.Api.Models
             /// <param name="writer">Serialization writer to use to serialize this model</param>
             public virtual void Serialize(ISerializationWriter writer)
             {
-                _ = writer ?? throw new ArgumentNullException(nameof(writer));
+                if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
                 if(Integer != null)
                 {
                     writer.WriteIntValue(null, Integer);
@@ -163,7 +163,7 @@ namespace ArchiwindRevitAddIn.Api.Models
             /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
             public static global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_draft_credits CreateFromDiscriminatorValue(IParseNode parseNode)
             {
-                _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+                if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
                 var mappingValue = parseNode.GetChildNode("")?.GetStringValue();
                 var result = new global::ArchiwindRevitAddIn.Api.Models.BillingPlan.BillingPlan_draft_credits();
                 if(parseNode.GetIntValue() is int integerValue)
@@ -190,7 +190,7 @@ namespace ArchiwindRevitAddIn.Api.Models
             /// <param name="writer">Serialization writer to use to serialize this model</param>
             public virtual void Serialize(ISerializationWriter writer)
             {
-                _ = writer ?? throw new ArgumentNullException(nameof(writer));
+                if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
                 if(Integer != null)
                 {
                     writer.WriteIntValue(null, Integer);