+++ ArchiWindRevitAddIn/Api/Models/CreateSimulationV1Params.cs
@@ -52,7 +52,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static global::ArchiwindRevitAddIn.Api.Models.CreateSimulationV1Params CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.Models.CreateSimulationV1Params();
         }
         /// <summary>
@@ -77,7 +77,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="writer">Serialization writer to use to serialize this model</param>
         public virtual void Serialize(ISerializationWriter writer)
         {
-            _ = writer ?? throw new ArgumentNullException(nameof(writer));
+            if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.Coordinates>("coords", Coords);
             writer.WriteGuidValue("model_id", ModelId);
             writer.WriteStringValue("name", Name);