+++ ArchiWindRevitAddIn/Api/Models/ModelV1_files.cs
@@ -60,7 +60,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static global::ArchiwindRevitAddIn.Api.Models.ModelV1_files CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.Models.ModelV1_files();
         }
         /// <summary>
@@ -83,7 +83,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="writer">Serialization writer to use to serialize this model</param>
         public virtual void Serialize(ISerializationWriter writer)
         {
-            _ = writer ?? throw new ArgumentNullException(nameof(writer));
+            if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.ModelV1_files_building>("building", Building);
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.ModelV1_files_surroundings>("surroundings", Surroundings);
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.ModelV1_files_terrain>("terrain", Terrain);