+++ ArchiWindRevitAddIn/Api/Models/User.cs
@@ -54,7 +54,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static global::ArchiwindRevitAddIn.Api.Models.User CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.Models.User();
         }
         /// <summary>
@@ -77,7 +77,7 @@ namespace ArchiwindRevitAddIn.Api.Models
         /// <param name="writer">Serialization writer to use to serialize this model</param>
         public virtual void Serialize(ISerializationWriter writer)
         {
-            _ = writer ?? throw new ArgumentNullException(nameof(writer));
+            if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.BillingPlan>("billing_plan", BillingPlan);
             writer.WriteStringValue("email", Email);
             writer.WriteStringValue("full_name", FullName);