+++ ArchiWindRevitAddIn/Api/V1/ModelsRequests/ModelsRequestBuilder.cs
@@ -76,7 +76,7 @@ namespace ArchiwindRevitAddIn.Api.V1.ModelsRequests
         public async Task<global::ArchiwindRevitAddIn.Api.Models.ModelV1> PostAsync(global::ArchiwindRevitAddIn.Api.Models.CreateModelParamsV1 body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default, CancellationToken cancellationToken = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = ToPostRequestInformation(body, requestConfiguration);
             var errorMapping = new Dictionary<string, ParsableFactory<IParsable>>
             {
@@ -99,7 +99,7 @@ namespace ArchiwindRevitAddIn.Api.V1.ModelsRequests
         public RequestInformation ToPostRequestInformation(global::ArchiwindRevitAddIn.Api.Models.CreateModelParamsV1 body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = new RequestInformation(Method.POST, UrlTemplate, PathParameters);
             requestInfo.Configure(requestConfiguration);
             requestInfo.Headers.TryAdd("Accept", "application/json");