+++ ArchiWindRevitAddIn/Api/V1/Projects/Item/ItemRequestBuilder.cs
@@ -105,7 +105,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Projects.Item
         public async Task<global::ArchiwindRevitAddIn.Api.Models.ProjectV1> PutAsync(global::ArchiwindRevitAddIn.Api.Models.UpdateProjectV1Params body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default, CancellationToken cancellationToken = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = ToPutRequestInformation(body, requestConfiguration);
             var errorMapping = new Dictionary<string, ParsableFactory<IParsable>>
             {
@@ -160,7 +160,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Projects.Item
         public RequestInformation ToPutRequestInformation(global::ArchiwindRevitAddIn.Api.Models.UpdateProjectV1Params body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = new RequestInformation(Method.PUT, UrlTemplate, PathParameters);
             requestInfo.Configure(requestConfiguration);
             requestInfo.Headers.TryAdd("Accept", "application/json");