+++ ArchiWindRevitAddIn/Api/V1/Projects/Item/Simulations/SimulationsGetResponse.cs
@@ -45,7 +45,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static global::ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations.SimulationsGetResponse CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations.SimulationsGetResponse();
         }
         /// <summary>
@@ -66,7 +66,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations
         /// <param name="writer">Serialization writer to use to serialize this model</param>
         public virtual void Serialize(ISerializationWriter writer)
         {
-            _ = writer ?? throw new ArgumentNullException(nameof(writer));
+            if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
             writer.WriteCollectionOfObjectValues<global::ArchiwindRevitAddIn.Api.Models.SimulationV1>("items", Items);
             writer.WriteObjectValue<global::ArchiwindRevitAddIn.Api.Models.ListPageNav>("nav", Nav);
             writer.WriteAdditionalData(AdditionalData);