+++ ArchiWindRevitAddIn/Api/V1/Projects/Item/Simulations/SimulationsResponse.cs
@@ -20,7 +20,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
         public static new global::ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations.SimulationsResponse CreateFromDiscriminatorValue(IParseNode parseNode)
         {
-            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
+            if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
             return new global::ArchiwindRevitAddIn.Api.V1.Projects.Item.Simulations.SimulationsResponse();
         }
     }