+++ ArchiWindRevitAddIn/Api/V1/Simulations/SimulationsRequestBuilder.cs
@@ -83,7 +83,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Simulations
         public async Task<global::ArchiwindRevitAddIn.Api.Models.SimulationV1> PostAsync(global::ArchiwindRevitAddIn.Api.Models.CreateSimulationV1Params body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default, CancellationToken cancellationToken = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = ToPostRequestInformation(body, requestConfiguration);
             var errorMapping = new Dictionary<string, ParsableFactory<IParsable>>
             {
@@ -107,7 +107,7 @@ namespace ArchiwindRevitAddIn.Api.V1.Simulations
         public RequestInformation ToPostRequestInformation(global::ArchiwindRevitAddIn.Api.Models.CreateSimulationV1Params body, Action<RequestConfiguration<DefaultQueryParameters>> requestConfiguration = default)
         {
 #endif
-            _ = body ?? throw new ArgumentNullException(nameof(body));
+            if(ReferenceEquals(body, null)) throw new ArgumentNullException(nameof(body));
             var requestInfo = new RequestInformation(Method.POST, UrlTemplate, PathParameters);
             requestInfo.Configure(requestConfiguration);
             requestInfo.Headers.TryAdd("Accept", "application/json");