+++ ArchiWindRevitAddIn/Application.cs
@@ -1,6 +1,5 @@
 ﻿using ArchiWindRevitAddIn.Services;
 using Autodesk.Revit.UI;
-using Nice3point.Revit.Extensions.UI;
 using Nice3point.Revit.Toolkit.External;
 
 namespace ArchiWindRevitAddIn